ubuntu/book_sales_website/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.0
Werkzeug==3.0.1
paypalrestsdk==1.13.1
reportlab==4.0.7
Pillow
